# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aimemory REST API（管理台自用面）",
    "version": "0.2.0",
    "description": "个人自托管 AI 记忆服务（mem0 形态）。对外 mem0 形态 API（/v1 /v2）与管理台自用面（/api）同库。鉴权：Authorization: Token m0-xxx（API key）或 aim_session cookie（Web 口令登录）；除标注「公开」外全部需要鉴权。请求/响应主体为 JSON；schemas 按需渐进补全，本契约当前承诺「路径与方法集」与实现一致（由 test/api-contract.test.js 守护）。",
    "x-data-model": "memories 事实记忆（mem0 语义：add 异步提炼 + 冲突消解 + 向量/FTS 混合检索 + memory_ops 变更历史）"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "system"
    },
    {
      "name": "memories"
    },
    {
      "name": "keys"
    }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "健康探针（公开）",
        "description": "db / llm / embedding 连通性"
      }
    },
    "/mcp": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "MCP Streamable HTTP 端点"
      },
      "get": {
        "tags": [
          "system"
        ],
        "summary": "MCP 端点信息"
      }
    },
    "/api/openapi.json": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "本契约（公开直出）"
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "当前身份"
      }
    },
    "/api/stats": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "记忆与密钥计数"
      }
    },
    "/api/memories": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "分页列出记忆",
        "description": "filters 支持 metadata 键值与 created_at/updated_at 时间范围"
      },
      "post": {
        "tags": [
          "memories"
        ],
        "summary": "受理素材（text/messages）",
        "description": "异步：返回 event_id，后台 LLM 提炼 + 冲突消解后入库；提炼失败不落库"
      }
    },
    "/api/memories/export": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "导出全量记忆（JSON）"
      }
    },
    "/api/memories/{id}": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "读取单条记忆"
      },
      "delete": {
        "tags": [
          "memories"
        ],
        "summary": "删除单条记忆"
      }
    },
    "/api/events/{id}": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "查询异步素材任务状态",
        "description": "status: pending|processing|done|failed；done 回执含 count/memories/ops（消解明细）"
      }
    },
    "/api/keys": {
      "get": {
        "tags": [
          "keys"
        ],
        "summary": "列出 API Token"
      },
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "签发 API Token（返回明文 m0-xxx）"
      }
    },
    "/api/keys/{id}/revoke": {
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "吊销 API Token"
      }
    },
    "/v1/memories": {
      "post": {
        "summary": "add：提交记忆素材（异步提炼）",
        "tags": [
          "memories"
        ],
        "description": "infer=true（默认）异步受理：返回 event_id，用 GET /v1/event/{event_id} 轮询；infer=false 原文直存同步返回 results。user_id 省略时默认鉴权主体；agent_id/run_id 为自由标签维度。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "按作用域批量删除（异步）",
        "tags": [
          "memories"
        ],
        "description": "query 参数 user_id / agent_id / run_id 至少其一（mem0 语义：'*' = 全选该维度）。返回 event_id。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/{id}": {
      "get": {
        "summary": "单条记忆",
        "tags": [
          "memories"
        ],
        "description": "mem0 形状：文本字段名为 memory。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新记忆（text/metadata）",
        "tags": [
          "memories"
        ],
        "description": "改文本后异步重消解；变更写入 history。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除单条记忆",
        "tags": [
          "memories"
        ],
        "description": "删除并写入 DELETE 历史。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories/{id}/history": {
      "get": {
        "summary": "记忆变更历史",
        "tags": [
          "memories"
        ],
        "description": "ADD/UPDATE/DELETE 每次变更留痕（new/old_memory），新→旧。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v2/memories": {
      "post": {
        "summary": "get_all：按 filters 分页列出",
        "tags": [
          "memories"
        ],
        "description": "body: {filters, page, page_size}；filters 至少含一个实体维度（user_id/agent_id/run_id，user_id 省略默认主体）。响应 {count, next, previous, results}。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/v2/memories/search": {
      "post": {
        "summary": "search：语义+关键词混合检索",
        "tags": [
          "memories"
        ],
        "description": "body: {query, filters, top_k, threshold}。响应 {results: [{id, memory, score, ...}]}。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/v1/event/{event_id}": {
      "get": {
        "summary": "事件状态轮询",
        "tags": [
          "memories"
        ],
        "description": "add（infer=true）与批量删除的异步事件状态：pending/processing/done/failed。",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "x-exempt-notes": "HTML 页面（/、/admin、/auth/login、/connect）与 /skill/* 静态资源不属于 REST 契约。"
}